Refactoring complete
- Include optimization
- CMake file refactoring
- Version Update
- Project structure refactoring
- Issue completed https://github.com/xaprier/qt-notepad/issues/1
