+++ .gitignore
@@ -3,7 +3,7 @@
 # jetbrains settings
 .idea
 # kdevelop settings
-src/.kdev4
+.kdev4
 # build dir
 build*
