+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.5)
 
-project(qt-notepad VERSION 0.1 LANGUAGES CXX)
+project(qt-notepad VERSION 0.2 LANGUAGES CXX)
 
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
 
@@ -10,31 +10,24 @@ set(CMAKE_AUTORCC ON)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-set(CMAKE_PREFIX_PATH "~/Qt/6.4.1/gcc_64/lib/cmake")
 
 find_package(QT NAMES Qt6 Qt5 REQUIRED COMPONENTS Widgets)
 find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Widgets)
 
+file(GLOB CPP_FILES src/*.cpp)
+file(GLOB UI_FILES design/*.ui)
+file(GLOB HPP_FILES include/*.hpp)
+file(GLOB QRC_FILES design/*.qrc)
+
 set(PROJECT_SOURCES
-    cpp-files/main.cpp
-
-    cpp-files/mainwindow.cpp
-    header-files/mainwindow.h
-    design/mainwindow.ui
-
-    cpp-files/finddialog.cpp
-    header-files/finddialog.h
-    design/finddialog.ui
-
-    cpp-files/replacedialog.cpp
-    header-files/replacedialog.h
-    design/replacedialog.ui
-
-    design/resources.qrc
+    ${CPP_FILES}
+    ${UI_FILES}
+    ${HPP_FILES}
+    ${QRC_FILES}
 )
 
 if(${QT_VERSION_MAJOR} GREATER_EQUAL 6)
-    qt_add_executable(qt-notepad
+    qt_add_executable(${PROJECT_NAME}
         MANUAL_FINALIZATION
         ${PROJECT_SOURCES}
     )
@@ -45,22 +38,24 @@ if(${QT_VERSION_MAJOR} GREATER_EQUAL 6)
 # For more information, see https://doc.qt.io/qt-6/qt-add-executable.html#target-creation
 else()
     if(ANDROID)
-        add_library(qt-notepad SHARED
+        add_library(${PROJECT_NAME} SHARED
             ${PROJECT_SOURCES}
         )
 
     # Define properties for Android with Qt 5 after find_package() calls as:
     # set(ANDROID_PACKAGE_SOURCE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/android")
     else()
-        add_executable(qt-notepad
+        add_executable(${PROJECT_NAME}
             ${PROJECT_SOURCES}
         )
     endif()
 endif()
 
-target_link_libraries(qt-notepad PRIVATE Qt${QT_VERSION_MAJOR}::Widgets)
+target_include_directories(${PROJECT_NAME} PRIVATE "./include")
 
-set_target_properties(qt-notepad PROPERTIES
+target_link_libraries(${PROJECT_NAME} PRIVATE Qt${QT_VERSION_MAJOR}::Widgets)
+
+set_target_properties(${PROJECT_NAME} PROPERTIES
     MACOSX_BUNDLE_GUI_IDENTIFIER my.example.com
     MACOSX_BUNDLE_BUNDLE_VERSION ${PROJECT_VERSION}
     MACOSX_BUNDLE_SHORT_VERSION_STRING ${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}
@@ -69,5 +64,5 @@ set_target_properties(qt-notepad PROPERTIES
 )
 
 if(QT_VERSION_MAJOR EQUAL 6)
-    qt_finalize_executable(qt-notepad)
+    qt_finalize_executable(${PROJECT_NAME})
 endif()
