+++ design/resources.qrc
