--- src/.kdev4/src.kdev4
@@ -1,19 +1,0 @@
-[Buildset]
-BuildItems=@Variant(\x00\x00\x00\t\x00\x00\x00\x00\x01\x00\x00\x00\x0b\x00\x00\x00\x00\x01\x00\x00\x00\x14\x00q\x00t\x00-\x00n\x00o\x00t\x00e\x00p\x00a\x00d)
-
-[CMake]
-Build Directory Count=1
-Current Build Directory Index-Host System=0
-
-[CMake][CMake Build Directory 0]
-Build Directory Path=/run/media/xaprier/Project/Projects/git/qt-notepad/build
-Build Type=Release
-CMake Binary=/usr/bin/cmake
-CMake Executable=/usr/bin/cmake
-Environment Profile=
-Extra Arguments=
-Install Directory=/usr/local
-Runtime=Host System
-
-[Project]
-VersionControlSupport=kdevgit
